+++ navigation/atlassian-switcher/src/create-custom-provider.ts
@@ -1,3 +1,4 @@
 export { createAvailableProductsProvider } from './common/providers/default-available-products-provider';
 export { createJoinableSitesProvider } from './cross-join/providers/default-joinable-sites-provider';
 export { defaultFetchData as defaultJoinableSitesFetch } from './cross-join/providers/default-joinable-sites-fetch';
+export { fetchProductRecommendations } from './cross-join/providers/product-recommendations-fetch';